display menu function added to another file from main.rb
