+++ main.rb
@@ -1,4 +1,5 @@
 require_relative 'app'
+require_relative 'display_menu'
 
 def main
   books = []
@@ -15,18 +16,6 @@ def main
 
     break if option == 7
   end
-end
-
-def display_menu
-  puts 'Library Management System'
-  puts '1. List all books'
-  puts '2. List all people'
-  puts '3. Create a person'
-  puts '4. Create a book'
-  puts '5. Create a rental'
-  puts '6. List rentals for a person'
-  puts '7. Quit'
-  print 'Select an option: '
 end
 
 def handle_option(option, books, people, rentals)
